# [js/webgpu] Optimize grouped conv