--- a/js/web/lib/wasm/jsep/webgpu/ops/conv.ts
+++ b/js/web/lib/wasm/jsep/webgpu/ops/conv.ts
@@ -9,7 +9,7 @@ import { ComputeContext } from '../types';
 import { createConv2DMatMulProgramInfo } from './3rd-party/conv2d_mm_webgpu';
 import { computeConv3DInfo, createConv3DNaiveProgramInfo } from './3rd-party/conv3d_naive_webgpu';
 import { createMatmulProgramInfo } from './3rd-party/matmul_packed_webgpu';
-import { createGroupedConvProgramInfo } from './conv-grouped';
+import { createGroupedConvProgramInfo, createGroupedConvVectorizeProgramInfo } from './conv-grouped';
 import { InternalActivationAttributes, parseInternalActivationAttributes } from './fuse-utils';
 import { createNaiveMatmulProgramInfo } from './matmul';
 import { createTransposeProgramInfo } from './transpose';
@@ -171,19 +171,8 @@ const conv2d = (
     isChannelsLast,
   );
   if (attributes.group !== 1) {
-    // NVIDIA GPU with ampere architecture fails with below 2 cases, but we couldn't repro them with any other
-    // GPUs. So just disable vectorize on NVIDIA ampere to ensure always correct outputs.
-    // [webgpu]Conv - conv - vectorize group - B
-    // [webgpu]Conv - conv - vectorize group - D
-    /*  const enableGroupedConvVectorize = !context.adapterInfo.isArchitecture('ampere');
-    if (
-      enableGroupedConvVectorize &&
-      isChannelsLast &&
-      inputs[1].dims[0] === attributes.group &&
-      inputs[1].dims[1] === 1 &&
-      attributes.dilations[0] === 1 &&
-      attributes.dilations[1] === 1
-    ) {
+    const convInputs = [inputs[0]];
+    if (isChannelsLast) {
       const transposedWeight =
         (context.kernelCustomData.wT as TensorView | undefined) ??
         context.compute(createTransposeProgramInfo(inputs[1], weightTransposeAttribute), {
@@ -193,33 +182,31 @@ const conv2d = (
       if (attributes.wIsConst && !context.kernelCustomData.wT) {
         context.kernelCustomData.wT = transposedWeight;
       }
-      const convInputs = [inputs[0], transposedWeight];
-      if (inputs.length === 3) {
-        convInputs.push(inputs[2]);
-      }
+      convInputs.push(transposedWeight);
+    } else {
+      convInputs.push(inputs[1]);
+    }
+    if (inputs.length === 3) {
+      convInputs.push(inputs[2]);
+    }
+    // NVIDIA GPU with ampere architecture fails with below 2 cases, but we couldn't repro them with any other
+    // GPUs. So just disable vectorize on NVIDIA ampere to ensure always correct outputs.
+    // [webgpu]Conv - conv - vectorize group - B
+    // [webgpu]Conv - conv - vectorize group - D
+    const enableGroupedConvVectorize = !context.adapterInfo.isArchitecture('ampere');
+    if (
+      enableGroupedConvVectorize &&
+      isChannelsLast &&
+      inputs[1].dims[0] === attributes.group &&
+      inputs[1].dims[1] === 1 &&
+      attributes.dilations[0] === 1 &&
+      attributes.dilations[1] === 1
+    ) {
       context.compute(
         createGroupedConvVectorizeProgramInfo(convInputs, attributes, outputShape, squeezeOutputShapeFunction),
         { inputs: convInputs },
       );
-    } else*/ {
-      const convInputs = [inputs[0]];
-      if (isChannelsLast) {
-        const transposedWeight =
-          (context.kernelCustomData.wT as TensorView | undefined) ??
-          context.compute(createTransposeProgramInfo(inputs[1], weightTransposeAttribute), {
-            inputs: [1],
-            outputs: [attributes.wIsConst ? -2 : -1],
-          })[0];
-        if (attributes.wIsConst && !context.kernelCustomData.wT) {
-          context.kernelCustomData.wT = transposedWeight;
-        }
-        convInputs.push(transposedWeight);
-      } else {
-        convInputs.push(inputs[1]);
-      }
-      if (inputs.length === 3) {
-        convInputs.push(inputs[2]);
-      }
+    } else {
       context.compute(createGroupedConvProgramInfo(convInputs, attributes, outputShape, squeezeOutputShapeFunction), {
         inputs: convInputs,
       });