… js/web/test/data/ops/conv.jsonc