@fs-eire 2024-09-03
/azp run Android CI Pipeline,iOS CI Pipeline,ONNX Runtime React Native CI Pipeline